move merchandise into module
